model backend documentation
