+++ docs/make.jl
@@ -17,7 +17,8 @@ makedocs(;
     pages = [
         "Home" => "index.md",
         "Signals: The Core of Reactivity" => "signals.md",
-        "Probabilistic Inference with Cortex" => "inference.md"
+        "Probabilistic Model Backend" => "model_backend.md",
+        "Probabilistic Inference with Cortex" => "inference_engine.md"
     ]
 )
